Cleanup
Testing is going to be a real mess. To make that easier, I'm going to start
moving a lot of the event handlers out into logical modules. Void methods are
being converted to event handlers so they can be externalized. Some methods
will be tacked on a la timerMixin.

Going forward, using event handlers instead of proper methods will result
in a bit of a perf hit, but a relatively minor one. We'll gain tons of
insights, however, as this will give us the ability to write developer tooling
that will allow you to hook into the cluster and inspect the state of the
quorum.